UnoCore: rename C++ header-file
