+++ lib/UnoCore/UnoCore.unoproj
@@ -16,7 +16,7 @@
     "Backends/CIL/CIL.uxl:Extensions",
     "Backends/CPlusPlus/CPlusPlus.uxl:Extensions",
     "Backends/CPlusPlus/Uno.cpp.uxl:Extensions",
-    "Backends/CPlusPlus/Uno.h:File",
+    "Backends/CPlusPlus/uno.h:File",
     "Backends/CPlusPlus/Uno/_config.h:File",
     "Backends/CPlusPlus/Uno/_internal.h:File",
     "Backends/CPlusPlus/Uno/_invoke.cpp:File",
